move fsms to own pass
